New approach.

+++ MongodbAccess/Services/MongodbDeleteRepository.cs
@@ -1,5 +1,4 @@
 ﻿using MongoDB.Driver;
-using MongodbAccess.Model;
 using MongodbAccess.Services;
 using RepositoryAccess;
 using System;
@@ -8,12 +7,17 @@ using System.Threading.Tasks;
 
 namespace MongodbAccess.Implementations
 {
-    public class MongodbDeleteRepository<T> : MongodbRepository<T>, IDeleteRepository<T> where T : IdObject
+    public class MongodbDeleteRepository<T> : MongodbRepository<T>, IDeleteRepository<T>
     {
         public MongodbDeleteRepository(IMongoDatabase mongoDatabase) : base(mongoDatabase) { }
 
         public async Task DeleteAllByConditionsAsync(Expression<Func<T, bool>> expression)
         {
+            if (expression == null)
+            {
+                throw new ArgumentNullException(nameof(expression));
+            }
+
             await this._mongoCollection.DeleteManyAsync(expression);
         }
     }
